스탯 관련 json파일 & 함수 추가

+++ Server/Server/Game/Object/Player.cs
@@ -24,7 +24,7 @@ namespace Server.Game
             ObjectType = GameObjectType.Player;
         }
 
-        public override void OnDamaged(GameObject attacker, int damage)
+        public override void OnDamaged(GameObject attacker, float damage)
         {
             base.OnDamaged(attacker, damage);
         }
